Use actual ECS task definition file, drop name req
Previously the task definition file required by this step was actually just
an array of container defintions -- one of three main sections of a task
definition. This commit reads the task definition file as a full ECS task
definition, including keys for the family name and the optional 'volumes'
section.  Since the family is now defined in this file, the need to specify
the family as an argument in this step has been eliminated, and so that
argument has been removed.